requirements.txt
# requirements.txt
# Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector store
chromadb>=0.4.0

# Document loaders
pypdf>=3.17.0
python-docx>=1.0.0
docx2txt>=0.8
odfpy>=1.4.1
unstructured>=0.11.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1  # Optionnel si vous gardez YAML

# Optional: pour .doc ancien format
# antiword (système)
